feat(purchase): add PurchaseSetup and product domain models
- Implement PurchaseSetup builder for purchase initialization
- Add ProductDefinition, DiscountRule, and related APIs
- Introduce PaywallContext for paywall UI configuration
- Extend View with .withPurchases() modifiers for setup
- Provide type-erased configs and marketing fluent API
- Ensure backwards compatibility with legacy aliases
